variables

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	float f;		/* A floating-point variable */
 
 	/* Print thesize of each data type along with a description */
-	printf("Size of a char: %lu byts(s)\n", sizeof(a));
-	printf("Size of an int: %lu bytes(s)\n", sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
-	printf("Size of a float: %lu bytes(s)\n", sizeof(f));
+	printf("Size of a char: %lu byts(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu bytes(s)\n",(unsigned long) sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n",(unsigned long) sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 	/* Return 0 to indicate successful program execution */
 	return (0);
 }
